refactor: auth context 파일 분리
